requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.7.14
charset-normalizer==3.4.2
cloudscraper==1.2.71
contourpy==1.3.2
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
fonttools==4.59.0
frozenlist==1.7.0
h11==0.16.0
idna==3.10
kiwisolver==1.4.8
markdown2==2.5.3
matplotlib==3.10.3
multidict==6.6.3
numpy==2.3.2
outcome==1.3.0.post0
packaging==25.0
pillow==10.4.0
propcache==0.3.2
psutil==7.0.0
pyparsing==3.2.3
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-magic==0.4.27
requests==2.32.4
requests-toolbelt==1.0.0
selenium==4.34.2
shiboken6==6.9.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
tkinterweb==4.4.4
tkinterweb-tkhtml==1.1.1
trio==0.30.0
trio-websocket==0.12.2
ttkbootstrap==1.14.1
typing_extensions==4.14.1
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.20.1
